remove all occurrences of DiskInfo
